Form Component exists, needs debugging
Removed obsolete components, and updated router-links.
